requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
gunicorn==20.1.0